+++ hls4ml/writer/quartus_writer.py
@@ -1330,6 +1330,8 @@ class QuartusWriter(Writer):
         Args:
             model (ModelGraph): the hls4ml model.
         """
+        if not self.should_write_tar(model):
+            return
         tar_path = model.config.get_output_dir() + '.tar.gz'
         if os.path.exists(tar_path):
             os.remove(tar_path)
@@ -1349,6 +1351,4 @@ class QuartusWriter(Writer):
         self.write_nnet_utils(model)
         self.write_activation_tables(model)
         self.write_yml(model)
-        write_tar = model.config.get_writer_config().get('WriteTar', False)
-        if write_tar:
-            self.write_tar(model)
+        self.write_tar(model)
