+++ hls4ml/writer/vivado_writer.py
@@ -1112,6 +1112,8 @@ class VivadoWriter(Writer):
         Args:
             model (ModelGraph): the hls4ml model.
         """
+        if not self.should_write_tar(model):
+            return
         tar_path = model.config.get_output_dir() + '.tar.gz'
         if os.path.exists(tar_path):
             os.remove(tar_path)
@@ -1132,9 +1134,7 @@ class VivadoWriter(Writer):
             self.write_nnet_utils(model)
             self.write_generated_code(model)
             self.write_yml(model)
-            write_tar = model.config.get_writer_config().get('WriteTar', False)
-            if write_tar:
-                self.write_tar(model)
+            self.write_tar(model)
         else:
             self.write_project_dir(model)
             self.write_build_script_multigraph(model)
